+++ repos/ktor/common/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/common/one_to_many/OneToManyRoutes.kt
@@ -13,6 +13,7 @@ const val onDataClearedRoute = "onDataCleared"
 
 const val addRoute = "add"
 const val removeRoute = "remove"
+const val removeWithValueRoute = "removeWithValue"
 const val clearRoute = "clear"
 const val clearWithValueRoute = "clearWithValue"
 const val setRoute = "set"